refactor: 폴더 및 flask app의 호출 방식 수정
- app/__init__.py의 내용을 app.py로 옮겨 직접 app.run을 수행하도록 했습니다.
- 그 이유는 host 0.0.0.0을 입력해 로컬의 모든 주소에서 접근하도록 하기 위해서입니다.
- app 폴더의 이름을 resources 로 수정했습니다.
- 데이터베이스 관련 설정을 database.py로 옮겼습니다.

resolves: #21
