+++ resources/models.py
@@ -1,4 +1,4 @@
-from app import db
+from database import db
 
 
 class Company(db.Model):
